finalizando módulo de ambiente e iniciando caronte

+++ backend/src/environments/environments.service.ts
@@ -1,8 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, BadRequestException, HttpException, HttpStatus } from '@nestjs/common';
 import { CreateEnvironmentDto } from './dto/create-environment.dto';
 import { UpdateEnvironmentDto } from './dto/update-environment.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateUserDto } from 'src/users/dto/create-user.dto';
+import { AddUserInEnvironmentDto } from './dto/add-user-environment.dto';
+import { User } from '@prisma/client';
 
 @Injectable()
 export class EnvironmentsService {
@@ -20,10 +22,43 @@ export class EnvironmentsService {
     return env
   }
 
-  async createAndAddUser(data, envId: number) {
-    const user = await this.prisma.user.create({ // TODO: descobrir como relacionar ambiente com user criado pelo id
-      data
+  async addUserInEnvironment(data: AddUserInEnvironmentDto) {
+    const user = await this.prisma.user.findUniqueOrThrow({
+      where: { id: data.userId },
     })
+    
+    if (user.role !== data.role) {
+      throw new HttpException('User role is different of role provided', HttpStatus.BAD_REQUEST);
+    }
+    
+    try {
+      if (user.role === 'ADMIN') {
+        await this.prisma.environment.update({
+          where: { id: data.envId },
+          data: {
+            admins: { connect: { id: data.userId } }
+          }
+        })
+      } else {
+        await this.prisma.environment.update({
+          where: { id: data.envId },
+          data: {
+            frequenters: { connect: { id: data.userId } }
+          }
+        })
+      }
+
+      return {
+        status: 201,
+        message: 'User successfully added.'
+      }
+    } catch (error) {
+      if (error.code === 'P2025') {
+        throw new HttpException("Record not found", HttpStatus.NOT_FOUND);
+      } else {
+        throw new HttpException("Can't add user", HttpStatus.FORBIDDEN);
+      }
+    }
   }
 
   async findAll() {
@@ -37,7 +72,11 @@ export class EnvironmentsService {
   }
 
   async findOne(id: number) {
-    return await this.prisma.environment.findFirst({
+    if (!id) {
+      throw new BadRequestException('Invalid Input. ID must be sent.');
+    }
+
+    return await this.prisma.environment.findFirstOrThrow({
       where: { id },
       include: {
         admins: true,
@@ -48,6 +87,21 @@ export class EnvironmentsService {
   }
 
   async update(id: number, updateEnvironmentDto: UpdateEnvironmentDto) {
+    if (!id) {
+      throw new BadRequestException('Invalid Input. ID must be sent.');
+    }
+
+    const validFields = ['name', 'description', 'adminId'];
+    const invalidFields = Object.keys(updateEnvironmentDto).filter(
+      field => !validFields.includes(field),
+    );
+
+    if (invalidFields.length > 0) {
+      throw new BadRequestException(
+        `Invalid fields provided: ${invalidFields.join(', ')}`,
+      );
+    }
+
     return await this.prisma.environment.update({
       data: updateEnvironmentDto,
       where: { id }
@@ -55,8 +109,12 @@ export class EnvironmentsService {
   }
 
   async remove(id: number) {
+    if (!id) {
+      throw new BadRequestException('Invalid Input. ID must be sent.');
+    }
+
     return await this.prisma.environment.delete({
       where: { id }
-    });;
+    });
   }
 }
